Renamed IRunnerResultAnalyzer.ts to RunnerResultAnalyzer.ts.
